API

Types

There are two types, Dataset and SpacegroupType, correspond to SpglibDataset and SpglibSpacegroupType, respectively. They store basic information of a symmetry search. The struct Cell is to contain input data of a symmetry search.

Spglib.Cell — Type

Cell(lattice, positions, types, magmoms=nothing)

The basic input data type of Spglib.

Lattice parameters lattice are given by a 3×3 matrix with floating point values, where 𝐚, 𝐛, and 𝐜 are given as rows, which results in the transpose of the definition for C-API. Fractional atomic positions positions are given by a N×3 matrix with floating point values, where N is the number of atoms. Numbers to distinguish atomic species types are given by a list of N integers. The collinear polarizations magmoms only work with get_symmetry and are given as a list of N floating point values.

Spglib.Dataset — Type

This represents SpglibDataset, see https://spglib.github.io/spglib/dataset.html#spglib-dataset.

Spglib.SpacegroupType — Type

This represents SpglibSpacegroupType, see https://spglib.github.io/spglib/api.html#spg-get-spacegroup-type.

Methods

Some methods are exported here. You can find their official documentation on this page.

Spglib.get_hall_number_from_symmetry — Function

get_hall_number_from_symmetry(rotation::AbstractArray{T,3}, translation::AbstractMatrix, num_operations::Integer, symprec=1e-5) where {T}

Obtain hall_number from the set of symmetry operations.

This is expected to work well for the set of symmetry operations whose distortion is small. The aim of making this feature is to find space-group-type for the set of symmetry operations given by the other source than spglib. Note that the definition of symprec is different from usual one, but is given in the fractional coordinates and so it should be small like 1e-5.

Spglib.get_dataset — Function

get_dataset(cell::Cell, symprec=1e-8)

Search symmetry operations of an input unit cell structure.

Spglib.get_spacegroup_type — Function

get_spacegroup_type(hall_number::Integer)

Translate Hall number to space group type information.

get_spacegroup_type(cell::Cell, symprec=1e-5)

Get SpacegroupType from cell.

Spglib.get_international — Function

get_international(cell::Cell, symprec=1e-8)

Return the space group type in Hermann–Mauguin (international) notation.

Spglib.get_schoenflies — Function

get_schoenflies(cell::Cell, symprec=1e-8)

Return the space group type in Schoenflies notation.

Spglib.standardize_cell — Function

standardize_cell(cell::Cell; to_primitive=false, no_idealize=false, symprec=1e-5)

Return standardized cell.

The standardized unit cell is generated from an input unit cell structure and its symmetry found by the symmetry search. The choice of the setting for each space group type is as explained for get_dataset.

Spglib.find_primitive — Function

find_primitive(cell::Cell, symprec=1e-5)

Find the primitive cell of an input unit cell.

This function is now a shortcut of standardize_cell with to_primitive = true and no_idealize = false.

Spglib.refine_cell — Function

refine_cell(cell::Cell, symprec=1e-5)

Return refined cell.

The standardized crystal structure is obtained from a non-standard crystal structure which may be slightly distorted within a symmetry recognition tolerance, or whose primitive vectors are differently chosen, etc. This function is now a shortcut of standardize_cell with to_primitive = false and no_idealize = false.

Spglib.niggli_reduce! — Function

niggli_reduce!(lattice::AbstractMatrix, symprec=1e-5)

Apply Niggli reduction to input basis vectors lattice and the reduced basis vectors are overwritten to lattice.

Spglib.delaunay_reduce! — Function

delaunay_reduce!(lattice::AbstractMatrix, symprec=1e-5)

Apply Delaunay reduction to input basis vectors lattice and the reduced basis vectors are overwritten to lattice.

Spglib.get_multiplicity — Function

get_multiplicity(cell::Cell, symprec=1e-8)

Return the exact number of symmetry operations. An error is thrown when it fails.

Spglib.get_ir_reciprocal_mesh — Function

get_ir_reciprocal_mesh(cell::Cell, mesh, is_shift=falses(3); is_time_reversal=true, symprec=1e-5)

Return k-points mesh and k-point map to the irreducible k-points.

Irreducible reciprocal grid points are searched from uniform mesh grid points specified by mesh and is_shift. mesh stores three integers. Reciprocal primitive vectors are divided by the number stored in mesh with (0,0,0) point centering. The centering can be shifted only half of one mesh by setting 1 or true for each is_shift element. If 0 or false is set for is_shift, it means there is no shift. This limitation of shifting enables the irreducible k-point search significantly faster when the mesh is very dense.

The reducible uniform grid points are returned in reduced coordinates as grid_address. A map between reducible and irreducible points are returned as grid_mapping_table as in the indices of grid_address. The number of the irreducible k-points are returned as the return value. The time reversal symmetry is imposed by setting is_time_reversal.

Spglib.get_version — Function

get_version()

Obtain the version number of spglib.

This is the mergence of spg_get_major_version, spg_get_minor_version, and spg_get_micro_version in its C-API.