Error

Spglib.SpglibReturnCode — Type

Represent various return codes from the Spglib library.

Spglib.get_error_code — Function

get_error_code()

Return an instance of the enumerated type SpglibReturnCode.

Spglib.get_error_message — Function

get_error_message(code::SpglibReturnCode)

Get the corresponding error message for a given error code.

Spglib.check_error — Function

check_error()

Check if an error has occurred and throw an exception if so.