Brillouin zone meshes

Spglib.BrillouinZoneMesh — Type

BrillouinZoneMesh(mesh, is_shift, ir_mapping_table, grid_address)

Represent the Brillouin zone mesh for searching (irreducible) reciprocal grid points from uniform mesh grid points specified by mesh and is_shift.

Reciprocal primitive vectors are divided by the number stored in mesh with $(0, 0, 0)$-centering. The center of grid mesh is shifted half of a grid spacing along the corresponding reciprocal axis by setting 1 or true to each is_shift element. If 0 or false is set to each is_shift element, there is no shift. This limitation of shifting enables the irreducible k-point search to be significantly faster when the mesh is very dense.

The reducible uniform grid points are stored in fractional coordinates as grid_address. A map between reducible and irreducible points is stored as ir_mapping_table in the indices of grid_address.

Warning

The mapping table's indices start from 1, not 0 as in Python or C.