Public API

Contents

Public API

Contents

Index

Public interface

Index

Spglib.Cell

Spglib.Dataset

Spglib.SpacegroupType

Spglib.basis_vectors

Spglib.delaunay_reduce

Spglib.find_primitive

Spglib.get_dataset

Spglib.get_dataset_with_hall_number

Spglib.get_hall_number_from_symmetry

Spglib.get_international

Spglib.get_ir_reciprocal_mesh

Spglib.get_multiplicity

Spglib.get_schoenflies

Spglib.get_spacegroup_number

Spglib.get_spacegroup_type

Spglib.get_symmetry

Spglib.get_symmetry_from_database

Spglib.get_version

Spglib.niggli_reduce

Spglib.refine_cell

Spglib.standardize_cell

Public interface

Types

There are two types, Dataset and SpacegroupType, correspond to SpglibDataset and SpglibSpacegroupType, respectively. They store basic information of a symmetry search. The struct Cell is to contain input data of a symmetry search.

In Spglib, basis vectors are represented by three column vectors, following the convention of spglib. Coordinates of an atomic point are represented as three fractional values relative to basis vectors. So when constructing a Cell:

The lattice can be a $3×3$ matrix with columns as basis vectors, or it can be a vector containing the three basis vectors. To get those basis vectors from a Cell, use basis_vectors.

The atomic positions can be a $3×N$ matrix, where $N$ denotes the number of atoms in a cell. Or it can be a vector of $N$ vectors, where each vector represents an atom.

The types variable corresponds to different atomic types.

Spglib.Cell — Type

Cell(lattice, positions, types, magmoms=zeros(length(types)))

The basic input data type of Spglib.

Lattice parameters lattice are given by a $3×3$ matrix with floating point values, where $𝐚$, $𝐛$, and $𝐜$ are given as columns. Fractional atomic positions positions are given by a vector of $N$ vectors with floating point values, where $N$ is the number of atoms. Numbers to distinguish atomic species types are given by a list of $N$ integers. The collinear polarizations magmoms only work with get_symmetry and are given as a list of $N$ floating point values, or a vector of vectors.

Spglib.Dataset — Type

Dataset(spacegroup_number, hall_number, international_symbol, hall_symbol, choice, transformation_matrix, origin_shift, n_operations, rotations, translations, n_atoms, wyckoffs, site_symmetry_symbols, equivalent_atoms, crystallographic_orbits, primitive_lattice, mapping_to_primitive, n_std_atoms, std_lattice, std_types, std_positions, std_rotation_matrix, std_mapping_to_primitive, pointgroup_symbol)

Represent SpglibDataset, see its official documentation.

Note

Fields crystallographic_orbits and primitive_lattice are added after spglib v1.15.0.

Spglib.SpacegroupType — Type

SpglibSpacegroupType(number, international_short, international_full, international, schoenflies, hall_symbol, choice, pointgroup_international, pointgroup_schoenflies, arithmetic_crystal_class_number, arithmetic_crystal_class_symbol)

Represent SpglibSpacegroupType, see its official documentation.

Methods

Some methods are exported here. You can find their official documentation on this page.

Spglib.basis_vectors — Function

basis_vectors(cell::Cell)

Return the three basis vectors from cell.

Spglib.get_symmetry — Function

get_symmetry(cell::Cell, symprec=1e-5)

Return the symmetry operations of a cell.

Spglib.get_hall_number_from_symmetry — Function

get_hall_number_from_symmetry(rotation::AbstractArray{T,3}, translation::AbstractMatrix, num_operations::Integer, symprec=1e-5) where {T}

Obtain hall_number from the set of symmetry operations.

This is expected to work well for the set of symmetry operations whose distortion is small. The aim of making this feature is to find space-group-type for the set of symmetry operations given by the other source than spglib. Note that the definition of symprec is different from usual one, but is given in the fractional coordinates and so it should be small like 1e-5.

Spglib.get_dataset — Function

get_dataset(cell::Cell, symprec=1e-5)

Search symmetry operations of an input unit cell structure.

Spglib.get_dataset_with_hall_number — Function

get_dataset_with_hall_number(cell::Cell, hall_number::Integer, symprec=1e-5)

Search symmetry operations of an input unit cell structure, using a given Hall number.

Spglib.get_spacegroup_type — Function

get_spacegroup_type(hall_number::Integer)

Translate Hall number to space group type information.

get_spacegroup_type(cell::Cell, symprec=1e-5)

Get SpacegroupType from cell.

Spglib.get_symmetry_from_database — Function

get_symmetry_from_database(hall_number)

Return the symmetry operations given a hall_number.

This function allows to directly access to the space group operations in the spglib database. To specify the space group type with a specific choice, hall_number is used.

Spglib.get_spacegroup_number — Function

get_spacegroup_number(cell::Cell, symprec=1e-5)

Get the spacegroup number of a cell.

Spglib.get_international — Function

get_international(cell::Cell, symprec=1e-5)

Return the space group type in Hermann–Mauguin (international) notation.

Spglib.get_schoenflies — Function

get_schoenflies(cell::Cell, symprec=1e-5)

Return the space group type in Schoenflies notation.

Spglib.standardize_cell — Function

standardize_cell(cell::Cell; to_primitive=false, no_idealize=false, symprec=1e-5)

Return standardized cell.

The standardized unit cell is generated from an input unit cell structure and its symmetry found by the symmetry search. The choice of the setting for each space group type is as explained for get_dataset.

Spglib.find_primitive — Function

find_primitive(cell::Cell, symprec=1e-5)

Find the primitive cell of an input unit cell.

This function is now a shortcut of standardize_cell with to_primitive = true and no_idealize = false.

Spglib.refine_cell — Function

refine_cell(cell::Cell, symprec=1e-5)

Return refined cell.

The standardized crystal structure is obtained from a non-standard crystal structure which may be slightly distorted within a symmetry recognition tolerance, or whose primitive vectors are differently chosen, etc. This function is now a shortcut of standardize_cell with to_primitive = false and no_idealize = false.

Spglib.niggli_reduce — Function

niggli_reduce(lattice::AbstractMatrix, symprec=1e-5)
niggli_reduce(cell::Cell, symprec=1e-5)

Apply Niggli reduction to input basis vectors lattice.

Spglib.delaunay_reduce — Function

delaunay_reduce(lattice::AbstractMatrix, symprec=1e-5)
delaunay_reduce(cell::Cell, symprec=1e-5)

Apply Delaunay reduction to input basis vectors lattice.

Spglib.get_multiplicity — Function

get_multiplicity(cell::Cell, symprec=1e-5)

Return the exact number of symmetry operations. An error is thrown when it fails.

Spglib.get_ir_reciprocal_mesh — Function

get_ir_reciprocal_mesh(cell::Cell, mesh, is_shift=falses(3); is_time_reversal=true, symprec=1e-5)

Return k-points mesh and k-point map to the irreducible k-points.

Irreducible reciprocal grid points are searched from uniform mesh grid points specified by mesh and is_shift. mesh stores three integers. Reciprocal primitive vectors are divided by the number stored in mesh with (0,0,0) point centering. The centering can be shifted only half of one mesh by setting 1 or true for each is_shift element. If 0 or false is set for is_shift, it means there is no shift. This limitation of shifting enables the irreducible k-point search significantly faster when the mesh is very dense.

The reducible uniform grid points are returned in reduced coordinates as grid_address. A map between reducible and irreducible points are returned as grid_mapping_table as in the indices of grid_address. The number of the irreducible k-points are returned as the return value. The time reversal symmetry is imposed by setting is_time_reversal.

Version 0.2

The returned mapping table is indexed starting at 1, not 0 as in Python or C.

Spglib.get_version — Function

get_version()

Obtain the version number of spglib.

This is the mergence of spg_get_major_version, spg_get_minor_version, and spg_get_micro_version in its C-API.